[lldb] Extract native process support
