+++ lldb/source/Plugins/Process/NetBSD/CMakeLists.txt
@@ -7,6 +7,7 @@ add_lldb_library(lldbPluginProcessNetBSD
   LINK_COMPONENTS
     Support
   LINK_LIBS
+    lldbNativeProcessSupport
     lldbHost
     lldbSymbol
     lldbTarget
