+++ lldb/source/Plugins/Process/Windows/Common/NativeRegisterContextWindows_arm64.h
@@ -12,7 +12,7 @@
 
 #include "NativeRegisterContextWindows.h"
 
-#include "Plugins/Process/Utility/NativeRegisterContextDBReg_arm64.h"
+#include "Plugins/Process/NativeCommon/NativeRegisterContextDBReg_arm64.h"
 #include "Register/Common/RegisterInfoCommon_arm64.h"
 #include "Register/Arch/lldb-arm64-register-enums.h"
