+++ lldb/source/Register/Arch/RegisterInfos_x86_64_with_base_shared.h
@@ -7,12 +7,11 @@
 //===----------------------------------------------------------------------===//
 
 #include "Register/Arch/lldb-x86-register-enums.h"
+#include "llvm/TargetParser/Triple.h"
 #include <stdint.h>
 
 #ifndef lldb_RegisterInfos_x86_64_with_base_shared_h
 #define lldb_RegisterInfos_x86_64_with_base_shared_h
-
-#include "Plugins/Process/Utility/NativeRegisterContextRegisterInfo.h"
 
 namespace lldb_private {
