+++ llvm/utils/gn/secondary/lldb/source/Plugins/Process/Utility/BUILD.gn
@@ -30,12 +30,6 @@ static_library("Utility") {
     "LinuxSignals.cpp",
     "MemoryTagManagerAArch64MTE.cpp",
     "NativeProcessSoftwareSingleStep.cpp",
-    "NativeRegisterContextDBReg.cpp",
-    "NativeRegisterContextDBReg_arm.cpp",
-    "NativeRegisterContextDBReg_arm64.cpp",
-    "NativeRegisterContextDBReg_loongarch.cpp",
-    "NativeRegisterContextDBReg_x86.cpp",
-    "NativeRegisterContextRegisterInfo.cpp",
     "NetBSDSignals.cpp",
     "OpenBSDSignals.cpp",
     "RegisterContextDarwin_arm.cpp",
